requirements.txt
python-dotenv
openai
pypdf
gradio 